streaming_platform/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-wtf==0.15.1
python-dotenv==0.19.0
werkzeug==2.0.1
moviepy==1.0.3
pillow==10.2.0
boto3==1.26.137
python-jose==3.3.0
requests==2.26.0 
SQLAlchemy==1.4.49